requirements.txt
openpyxl==3.0.9
pandas==1.4.2
mysql-connector-python==8.0.27
protobuf==3.19.3
python_dateutil==2.8.2
pytz==2021.3
setuptools==60.10.0
six==1.16.0
